pyproject.toml
[project]
name = "agentllm-agents-demo"
version = "1.0.0"
description = "Demo agents for AgentLLM"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "agentllm-core>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.entry-points."agentllm.agents"]
demo-agent = "agentllm_agents_demo.demo_agent:DemoAgentFactory"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentllm_agents_demo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
